refactor(de/aniworld): Move AniWorld extension to lib-multisrc
