+++ src/de/aniworld/build.gradle
@@ -1,13 +1,8 @@
 ext {
     extName = 'AniWorld'
     extClass = '.AniWorld'
-    extVersionCode = 31
+    themePkg = "aniworldtheme"
+    overrideVersionCode = 31
 }
 
 apply from: "$rootDir/common.gradle"
-
-dependencies {
-    implementation(project(':lib:voe-extractor'))
-    implementation(project(':lib:streamtape-extractor'))
-    implementation(project(':lib:dood-extractor'))
-}
